[Chore] 샘플 연습 추가 #102
